requirements.txt
# ============================================
# requirements.txt (ワークフロー機能 + OpenAI対応)
# ============================================

# 既存の必須依存関係
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=1.5.0

# OpenAI API
openai>=1.3.0 # Updated to a more recent version for client.chat.completions
tiktoken>=0.5.0 # For OpenAI token counting

# ワークフロー機能で使用する追加依存関係
asyncio-timeout>=4.0.2
python-dateutil>=2.8.2
jsonschema>=4.17.0 # For potential validation tasks
colorlog>=6.7.0 # For enhanced logging if used
pyarrow>=12.0.0 # For efficient data handling with pandas

# オプション: 開発時のデバッグとテスト用
rich>=13.0.0
pytest>=7.4.0
flake8>=6.0.0
mypy>=1.5.0

# 推奨: セキュリティとパフォーマンス向上
cryptography>=41.0.0
httpx>=0.24.0 # Streamlit/OpenAI/Google libs might use this internally
psutil>=5.9.0 # For memory monitoring (optional feature)

# UI/UX改善用の追加ライブラリ
plotly>=5.15.0 # For advanced visualizations (if visualization_tab uses it)
openpyxl>=3.1.0 # For Excel export/import if added
# polars>=0.18.0 # If pandas is replaced or supplemented with polars

# 🆕 Phase 1 で追加 (ワークフローのYAML定義用)
PyYAML>=6.0

# ============================================
# インストール手順
# ============================================
# 基本インストール（必須 + OpenAI）:
# pip install streamlit google-generativeai pandas openai tiktoken

# 完全インストール（全機能）:
# pip install -r requirements.txt